Fix delete item in uploading stage

+++ Yashmak.Web/Controllers/UploadController.cs
@@ -35,7 +35,8 @@
             var username = this.User.Identity.Name;
             var savePath = this.Server.MapPath(string.Format(@"~/App_Data/{0}/", username));
             var userId = this.User.Identity.GetUserId();
-            var userFilesInCurrentDir = this.repository.All().Where(d => d.ParentId == filenodeid && d.UserId == userId);
+            var userFilesInCurrentDir =
+                this.repository.All().Where(d => d.ParentId == filenodeid && d.UserId == userId);
 
             var statuses = new List<ViewDataUploadFileResult>();
 
@@ -44,26 +45,26 @@
                 var curFile = i;
                 var curFileName = this.Request.Files[curFile].FileName;
                 var saveFileName = Guid.NewGuid().ToString();
-                var st = FileSaver.StoreFile(x =>
-                    {
-                        x.File = this.Request.Files[curFile];
+                var st = FileSaver.StoreFile(
+                    x =>
+                        {
+                            x.File = this.Request.Files[curFile];
 
-                        // note how we are adding an additional value to be posted with delete request
-                        // and giving it the same value posted with upload
-                        x.DeleteUrl = this.Url.Action("DeleteFile", new { filenodeid = filenodeid });
-                        x.StorageDirectory = savePath;
-                        x.UrlPrefix = string.Format("/App_Data/{0}", username);
+                            // note how we are adding an additional value to be posted with delete request
+                            // and giving it the same value posted with upload
+                            x.StorageDirectory = savePath;
 
-                        // this is used to generate the relative url of the file
+                            // TODO: CHange the download url
+                            x.UrlPrefix = string.Format("/App_Data/{0}", username);
 
-                        // overriding defaults
-                        x.FileName = saveFileName;
+                            // overriding defaults
+                            x.FileName = saveFileName;
 
-                        // default is filename suffixed with filetimestamp
-                        x.ThrowExceptions = true;
+                            // default is filename suffixed with filetimestamp
+                            x.ThrowExceptions = true;
 
-                        // default is false, if false exception message is set in error property
-                    });
+                            // default is false, if false exception message is set in error property
+                        });
 
                 if (userFilesInCurrentDir.Any(f => f.FileName == curFileName))
                 {
@@ -75,44 +76,53 @@
                 }
                 else
                 {
-                    this.repository.Add(new File { FileName = curFileName, IsDirectory = false, ModifiedOn = DateTime.Now, ParentId = filenodeid, UserId = userId, Size = st.size, PathToFile = st.SavedFileName });
+                    var file = new File
+                        {
+                            FileName = curFileName, 
+                            IsDirectory = false, 
+                            ModifiedOn = DateTime.Now, 
+                            ParentId = filenodeid, 
+                            UserId = userId, 
+                            Size = st.size, 
+                            PathToFile = st.SavedFileName
+                        };
+                    this.repository.Add(file);
                     this.repository.SaveChanges();
+                    st.deleteUrl = this.Url.Action(
+                        "DeleteFile", 
+                        new { filenodeid = file.Id, filePath = st.FullPath });
                 }
 
                 statuses.Add(st);
             }
 
+            // TODO: fix downloading and thumbnail
             // statuses.ForEach(x => x.thumbnailUrl = x.url + "?width=80&height=80");
             // statuses.ForEach(x => x.url = this.Url.Action("DownloadFile", new { fileUrl = x.url, mimetype = x.type }));
             var viewresult = this.Json(new { files = statuses });
-
-            // for IE8 which does not accept application/json
-            if (this.Request.Headers["Accept"] != null && !this.Request.Headers["Accept"].Contains("application/json"))
-            {
-                viewresult.ContentType = "text/plain";
-            }
 
             return viewresult;
         }
 
         // here i am receving the extra info injected
         [HttpPost] // should accept only post
-        public ActionResult DeleteFile(int? filenodeid, string fileUrl)
+        public ActionResult DeleteFile(int? filenodeid, string filePath)
         {
-            var filePath = this.Server.MapPath("~" + fileUrl);
+            var file = this.repository.GetById((int)filenodeid);
+            if (file.UserId != this.User.Identity.GetUserId())
+            {
+                return
+                    this.Json(new { error = "Don't try delete files that do not belongs to you!" });
+            }
 
             if (System.IO.File.Exists(filePath))
             {
                 System.IO.File.Delete(filePath);
             }
 
-            var viewresult = this.Json(new { error = string.Empty });
+            this.repository.Delete(file);
 
-            // for IE8 which does not accept application/json
-            if (this.Request.Headers["Accept"] != null && !this.Request.Headers["Accept"].Contains("application/json"))
-            {
-                viewresult.ContentType = "text/plain";
-            }
+            var viewresult = this.Json(new { error = string.Empty });
 
             return viewresult; // trigger success
         }
